+++ CHANGELOG.md
@@ -5,3 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2018-07-32
+### Added
+- Changelog File
+- MIT License
+- Readme Documentation
+- Functionality to read node argv arguments into a complex object. Consisting of numeric, boolean, array parsing to their primitives
+- Configure node-args for enabling/disabling numeric conversion
+- Configure node-args for enabling/disabling array parsing
+- Configure node-args for enabling/disabling boolean parsing
+- Configure node-args for a specific delimiter when array parsing
+- Configure node-args for setting which string representation is to be seen as a boolean true/false value